--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/sil-org/nodeping-cli/cmd"
-
-func main() {
-	cmd.Execute()
-}
