+++ period_test.go
@@ -1,4 +1,4 @@
-package lib
+package nodeping
 
 import (
 	"testing"
